added code on residual diagnostics in LR
Residual diagnostics are used to check for assumptions of LR and often if we could do a transformation when assumptions are violated
Residual plot of residuals vs fitted values, residuals vs predictors are useful to check if residuals have a constant variance.
Residual plot vs index checks ordering (time, space) of data and lets us know if the ordering are independent. 
Histogram, QQ plot of residuals tell us that if residuals are normally distributed or not (skewed, fat-tailed, thin-tailed, etc).